Refactored go and mailhog to be more intelligent

+++ app/vm/cookbooks/mailhog/recipes/default.rb
@@ -3,11 +3,13 @@
 # Copyright:: 2020, Steve Kukla, All Rights Reserved.
 # frozen_string_literal: true
 
+go_install_path = node[:mailhog][:go_install_path]
 sendmail_path = node[:mailhog][:sendmail_path]
 service_file = node[:mailhog][:service_file]
 
 golang 'Install golang' do
 	action :install
+	not_if { ::Dir.exist?(go_install_path) }
 end
 
 mailhog 'Stop mailhog' do
@@ -17,6 +19,7 @@ end
 
 mailhog 'Install, configure, enable, and reload mailhog' do
 	action %i[install configure enable reload]
+	not_if { ::File.exist?(service_file) }
 end
 
 php 'Configure mailhog sendmail path and restart PHP' do
